+++ tests/integration/adapter/incremental/test_incremental_unique_key.py
@@ -35,6 +35,8 @@ union all
 select 'NY','New York','Manhattan',toDate('2021-04-01')
 union all
 select 'PA','Philadelphia','Philadelphia',toDate('2021-05-21')
+union all
+select 'CO','Denver','',toDate('2021-06-18')
 """
 
 
@@ -60,7 +62,8 @@ union all
 select 'NY','New York','Manhattan',toDate('2021-04-01')
 union all
 select 'PA','Philadelphia','Philadelphia',toDate('2021-05-21')
-
+union all
+select 'CO','Denver','',toDate('2021-06-18')
 """